pyproject.toml
[tool.poetry]
name = "auxil"
version = "0.1.0"
description = []
authors = []
license = "AGPL-3.0+"
[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.2.2"

[tool.poetry.dev-dependencies]
mypy = "^0.711"
pre-commit = "^1.17"
python-dotenv = "^0.10.3"

[tool.poetry.scripts]
start = "launcher:main"
dev = "launcher:dev"
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

